Move authentication and database connection classes to utility package
