+++ src/main/java/com/example/pupsis_main_dashboard/Test.java
@@ -24,7 +24,7 @@ public class Test extends Application {
 
         try {
             Stage initializedStage = utility.loadStage(
-                    "fxml/StudentDashboard.fxml",
+                    "fxml/FacultyDashboard.fxml",
                     "PUPSIS",
                     Objects.requireNonNull(getClass().getResource("/com/example/pupsis_main_dashboard/Images/PUPSJ Logo.png")).toExternalForm(),
                     StageAndSceneUtils.WindowSize.MEDIUM
