+++ src/main/java/com/example/pupsis_main_dashboard/controllers/EditPageController.java
@@ -1,8 +1,6 @@
 package com.example.pupsis_main_dashboard.controllers;
 
 import com.example.pupsis_main_dashboard.utility.Student;
-import com.example.pupsis_main_dashboard.utility.Subject;
-import javafx.application.Application;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.control.TableView;
@@ -12,20 +10,13 @@ import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.util.converter.DefaultStringConverter;
 import java.net.URL;
 import java.util.ResourceBundle;
-import com.example.pupsis_main_dashboard.databaseOperations.dbConnection2;
+import com.example.pupsis_main_dashboard.utility.DBConnection;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Parent;
-import javafx.scene.Scene;
 import javafx.scene.control.Alert;
 import javafx.scene.control.MenuButton;
-import javafx.stage.Stage;
 
-import java.io.IOException;
-import java.net.URL;
 import java.sql.*;
-import java.util.ResourceBundle;
 
 public class EditPageController implements Initializable {
     @FXML
@@ -57,7 +48,7 @@ public class EditPageController implements Initializable {
 
     private ObservableList<Student> studentsList = FXCollections.observableArrayList();
 
-    dbConnection2 dbConn = new dbConnection2();
+    DBConnection dbConn = new DBConnection();
     String url = dbConn.URL;
     String user = dbConn.USER;
     String password = dbConn.PASSWORD;
