+++ src/main/java/com/example/pupsis_main_dashboard/controllers/StudentLoginController.java
@@ -1,7 +1,7 @@
 package com.example.pupsis_main_dashboard.controllers;
 
 //import com.example.pupsis_main_dashboard.databaseOperations.PasswordHandler;
-import com.example.pupsis_main_dashboard.databaseOperations.DBConnection;
+import com.example.pupsis_main_dashboard.utility.DBConnection;
 import com.example.pupsis_main_dashboard.utility.EmailService;
 import com.example.pupsis_main_dashboard.utility.RememberMeHandler;
 import com.example.pupsis_main_dashboard.utility.StageAndSceneUtils;
@@ -14,10 +14,8 @@ import javafx.application.Platform;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
-import javafx.fxml.FXMLLoader;
 import javafx.geometry.Pos;
 import javafx.scene.Node;
-import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.scene.control.*;
 import javafx.scene.effect.GaussianBlur;
@@ -30,7 +28,6 @@ import javafx.scene.layout.StackPane;
 import javafx.scene.layout.VBox;
 import javafx.scene.paint.Color;
 import javafx.scene.shape.Circle;
-import javafx.stage.Modality;
 import javafx.stage.Stage;
 import javafx.util.Duration;
 
@@ -45,12 +42,11 @@ import java.util.Random;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.prefs.Preferences;
-import java.util.Objects;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import static com.example.pupsis_main_dashboard.auth.AuthenticationService.authenticate;
+import static com.example.pupsis_main_dashboard.utility.AuthenticationService.authenticate;
 //import static com.example.pupsis_main_dashboard.utility.AuthenticationService.authenticate;
 
 public class StudentLoginController {
