+++ src/main/java/com/example/pupsis_main_dashboard/utility/NotificationManager.java
@@ -1,19 +1,14 @@
 package com.example.pupsis_main_dashboard.utility;
 
-import com.example.pupsis_main_dashboard.databaseOperations.DBConnection;
 import javafx.application.Platform;
 import javafx.geometry.Pos;
 import javafx.scene.Scene;
 import javafx.scene.control.Button;
 import javafx.scene.control.Label;
-import javafx.scene.layout.HBox;
 import javafx.scene.layout.VBox;
 import javafx.stage.Modality;
-import javafx.stage.Popup;
 import javafx.stage.Stage;
 import javafx.stage.StageStyle;
-import javafx.util.Duration;
-import javafx.animation.PauseTransition;
 
 import javax.mail.MessagingException;
 import java.sql.Connection;
